requirements.txt
# Core ML and Data Processing
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# API and Web Framework
fastapi==0.100.1
uvicorn==0.23.2
streamlit==1.25.0
pydantic==2.1.1

# Data Streaming
kafka-python==2.0.2
redis==4.6.0

# Workflow Orchestration
apache-airflow==2.6.3

# AWS SDK
boto3==1.28.25
botocore==1.31.25

# Utilities
python-dotenv==1.0.0
requests==2.31.0
geopy==2.3.0
folium==0.14.0

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0
